requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
pillow==10.2.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
